chonk: remove slider.chonk
